pyproject.toml
[project]
name = "cli-data-processor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
cli-data-processor = "cli_data_processor.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
